fixed startup problem, which occured while steady_clock.now().time_since_epoch() < burstSize * timePerToken
At startup, the bucket should contain burstSize tokens. As the internal time_ was initialized with 0, this condition was not met as long as the value of steady_clock.now().time_since_epoch() (which possibly corresponds to the uptime of the system) was too low.

This patch uses an int64 rather than a uint64 to store the time. The reduced time resolution results in a timer wrap-around after 146235 rather than 292471 years, which should be acceptable.